more warnings removed

+++ inc/util/OverlapUtils.hpp
@@ -12,7 +12,7 @@ public:
 		Position const& rPosition2,
 		Size const& rSize2);
 private:
-	static bool OverlappingOnAxis_(int x1, int x2, int width1, int width2);
+	static bool OverlappingOnAxis_(double x1, double x2, int width1, int width2);
 };
 
 #endif
